requirements.txt
requests>=2.31.0
openai>=1.12.0
google-generativeai>=0.3.0
anthropic>=0.8.0